La PR #157 a été mergée, et main a reçu depuis la PR #158
(codex/evaluer-le-plan-de-routage-existant) qui traite le même sujet que mes
trois derniers commits : plan des niveaux de diffusion, activation du routage,
distribution des rapports. Convergence indépendante, y compris sur les
conclusions du plan — trois axes distincts, diffusion non ordinale, SSE hors
périmètre.

Arbitrages de fusion, en gardant la meilleure version de chaque côté :

- Branchement du routage : celui de main. Il est plus intégré (métriques de
  routage remontées dans le journal d'activité), et il apporte une contrainte
  d'idempotence, un cron d'escalade et un test unitaire que je n'avais pas.
- acknowledge() : version de main, bornée par communauté ET par contexte, avec
  paramètres obligatoires — plus stricte que la mienne, qui les rendait
  optionnels.
- notification_sent : les deux côtés avaient corrigé le drapeau mensonger. Je
  garde le code de main et mon commentaire, qui explique le pourquoi.
- Document de plan : celui de main fait foi. Le mien n'y ajoute qu'une section
  de compléments, mise en regard de son § 5 « risques à traiter ».

Mes ajouts sont reposés par-dessus, aucun n'existant dans main :
- écran de gestion des règles, sans lequel le moteur tourne à vide ;
- émission réelle des notifications ;
- route de relève GET /api/atak/notifications, sans laquelle les notifications
  écrites n'étaient lisibles par personne ;
- listForReport(), markNotified(), cloisonnement de findById().

Le service est refactoré en conséquence : il ne route plus lui-même, il relit
l'historique de diffusion pour savoir qui notifier. Ce qui a réellement été
écrit prime sur ce que l'appelant croyait demander. Une garde d'idempotence
empêche de renotifier une diffusion déjà annoncée.

+++ migrations/20260417103000_login_security_otp_legal_identity.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS `user_legal_identities` (
-  `id` bigint unsigned NOT NULL AUTO_INCREMENT,
-  `tenant_id` bigint unsigned NOT NULL,
-  `user_id` bigint unsigned NOT NULL,
+  `id` int unsigned NOT NULL AUTO_INCREMENT,
+  `tenant_id` int unsigned NOT NULL,
+  `user_id` int unsigned NOT NULL,
   `first_name` varchar(100) DEFAULT NULL,
   `last_name` varchar(100) DEFAULT NULL,
   `phone` varchar(50) DEFAULT NULL,
